XWIKI-9516: Create a new Wiki API that replaces xwiki-platform-wiki-mananager and xwiki-platform-workspace
* Bug fix: canDeleteWiki() was testing if the current script had programming rights, which is not useful

+++ xwiki-platform-core/xwiki-platform-wiki/xwiki-platform-wiki-script/src/main/java/org/xwiki/wiki/script/WikiManagerScriptService.java
@@ -158,13 +158,22 @@ public class WikiManagerScriptService implements ScriptService
      */
     public boolean deleteWiki(String wikiId)
     {
-        // Test right
+        // Test if the script has the programming right
         XWikiContext context = xcontextProvider.get();
+        if (!authorizationManager.hasAccess(Right.PROGRAM,
+                context.getDoc().getAuthorReference(), context.getDoc().getDocumentReference())) {
+            String errorMessage = "This script does not have the right to delete a wiki.";
+            error(errorMessage, new Exception(errorMessage));
+            return false;
+        }
+
+        // Test right
         if (!canDeleteWiki(context.getUser(), wikiId)) {
             String errorMessage = "You don't have the right to delete the wiki";
             error(errorMessage, new Exception(errorMessage));
             return false;
         }
+
         // Delete the wiki
         try {
             wikiManager.delete(wikiId);
@@ -187,11 +196,6 @@ public class WikiManagerScriptService implements ScriptService
     {
         String errorMessage = String.format("Error while getting the descriptor of wiki [%s]", wikiId);
         try {
-            // Check if the current script has the programing rights
-            XWikiContext context = xcontextProvider.get();
-            authorizationManager.checkAccess(Right.PROGRAM, context.getDoc().getAuthorReference(),
-                    context.getDoc().getDocumentReference());
-
             // Get the wiki owner
             WikiDescriptor descriptor = wikiDescriptorManager.getById(wikiId);
             if (descriptor == null) {
@@ -210,8 +214,6 @@ public class WikiManagerScriptService implements ScriptService
                 return true;
             }
         } catch (WikiManagerException e) {
-            error(errorMessage, e);
-        } catch (AccessDeniedException e) {
             error(errorMessage, e);
         }
